move errors from index.rs -> mod.rs
