+++ crates/runtime/src/search/full_text/mod.rs
@@ -13,7 +13,51 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+use arrow_schema::DataType;
+use snafu::Snafu;
+use tantivy::TantivyError;
+
+use crate::datafusion::Error as DataFusionError;
+
 pub mod analyzer_rule;
 pub mod connector;
 pub mod index;
 pub mod udtf;
+
+#[derive(Debug, Snafu)]
+#[snafu(visibility(pub))]
+pub enum Error {
+    #[snafu(display("Full text search requires a primary key, and the table did not have one.",))]
+    NoPrimaryKey,
+
+    #[snafu(display(
+        "Primary key column '{column}' used in search index has unsupported data type: '{data_type}'",
+    ))]
+    PrimaryKeyInvalidType { column: String, data_type: DataType },
+
+    #[snafu(display("Primary key column '{column}' not found in table.",))]
+    PrimaryKeyNotFound { column: String },
+
+    #[snafu(display("Failed to create a full text search index: {source}.",))]
+    IndexCreationError { source: TantivyError },
+
+    #[snafu(display("Failed to insert or update data into a full text search index: {source}.",))]
+    IndexInsertionError { source: TantivyError },
+
+    #[snafu(display("Failed to retrieve the data from the full text search index: {source}.",))]
+    FailedToRetrieveDataFromIndex { source: TantivyError },
+
+    #[snafu(display("Failed to retrieve the data from the underlying table: {source}.",))]
+    FailedToRetrieveDataFromSource { source: DataFusionError },
+
+    #[snafu(display("Failed to insert data into the full text search index: {source}.",))]
+    FailedToInsertDataIntoIndex { source: TantivyError },
+
+    #[snafu(display(
+        "Failed to create the full text search index. Context: {context}. Error: {source}.",
+    ))]
+    InvalidIndexingError {
+        source: Box<dyn std::error::Error + Send + Sync>,
+        context: String,
+    },
+}
